+++ src/client/components/account_roles_panel/account_roles_panel.js
@@ -5,6 +5,7 @@
  * - item: a ReactiveVar which holds the account object to edit (may be empty, but not null)
  * - isNew: true|false
  * - checker: a ReactiveVar which holds the parent Checker
+ * - amInstance: a ReactiveVar which holds the amClass instance
  */
 
 import './account_roles_panel.html';